created grid view
